chore: renames test directories to use __ naming convention
