+++ packages/osc-academic-hub/__mocks__/index.js
