+++ packages/osc-academic-hub/package.json
@@ -12,9 +12,9 @@
     "build:remix": "cross-env NODE_ENV=production remix build",
     "dev": "cross-env NODE_ENV=development npm run build:css && run-p dev:*",
     "dev:css": "PORT=3001 LOAD_PATH=app/styles PATH_TO_MAIN_SCSS=/app/styles/main.scss PATH_TO_DEST_MAIN_CSS=/app/styles/dest/main.css node bin/postcss.js",
-    "dev:remix": "cross-env PORT=3000 NODE_ENV=development binode --require ./mocks -- @remix-run/dev:remix dev",
+    "dev:remix": "cross-env PORT=3000 NODE_ENV=development binode --require ./__mocks__ -- @remix-run/dev:remix dev",
     "start": "remix-serve build",
-    "start:mocks": "binode --require ./mocks -- @remix-run/serve:remix-serve build",
+    "start:mocks": "binode --require ./__mocks__ -- @remix-run/serve:remix-serve build",
     "test": "vitest --coverage",
     "test:dev": "vitest run",
     "test:e2e:dev": "start-server-and-test dev http://localhost:3000 \"PORT=3000 npx playwright test\"",
