+++ packages/osc-academic-hub/vitest.config.ts
@@ -10,7 +10,7 @@ export default defineConfig({
     test: {
         globals: true,
         environment: 'happy-dom',
-        setupFiles: ['./test/setup-test-env.ts'],
+        setupFiles: ['./__test__/setup-test-env.ts'],
         exclude: ['./e2e/**/*']
     }
 });
