+++ packages/osc-ui/vite.config.ts
@@ -77,7 +77,7 @@ export default defineConfig({
         reporters: ['verbose'],
         globals: true,
         environment: 'happy-dom',
-        setupFiles: ['./test/setup-test-env.ts'],
+        setupFiles: ['./__test__/setup-test-env.ts'],
         include: ['./src/**/*.{test,spec}.{js,mjs,cjs,ts,mts,cts,jsx,tsx,css}']
     }
 });
